feat: finish restructure and api calls for i_types, hausgeld and overview

+++ BE.Application/ImmobilienTypes/Commands/DeleteTypes/DeleteImmobilienTypesByIdCommandHandler.cs
@@ -16,12 +16,17 @@ namespace BE.Application.ImmobilienTypes.Commands.DeleteTypes
         {
             logger.LogInformation("Deleting ImmobilienType with {Id} {Type}", request.Id, request);
 
-
             var type = await typeRepository.GetByIdAsync(request.Id);
+            var isInUse = await typeRepository.IsInUseAsync(request.Id);
 
             if (type is null)
             {
                 throw new NotFoundException(nameof(ImmobilienType), request.Id.ToString());
+            }
+
+            if (isInUse)
+            {
+                throw new InvalidOperationException($"The ImmobilienType with Id: {request.Id} is still in use and cannot be deleted.");
             }
 
             await typeRepository.Delete(type);
